Feat: Implement Core Configurations and server-side architecture issue(#11)
